requirements.txt
streamlit
google.generativeai as genai
pypdf
Pillow
opencv-python
python-docx
opencv-python-headless